+++ namerd/storage/consul/src/test/scala/io/buoyant/namerd/storage/consul/ConsulConfigTest.scala
@@ -17,13 +17,15 @@ class ConsulConfigTest extends FunSuite with OptionValues {
          |experimental: true
          |pathPrefix: /foo/bar
          |host: consul.local
+         |token: some-token
          |port: 80
       """.stripMargin
     val mapper = Parser.objectMapper(yaml, Iterable(Seq(ConsulDtabStoreInitializer)))
-    val etcd = mapper.readValue[DtabStoreConfig](yaml).asInstanceOf[ConsulConfig]
-    assert(etcd.host.value == "consul.local")
-    assert(etcd.port.value == Port(80))
-    assert(etcd.pathPrefix == Some(Path.read("/foo/bar")))
+    val consul = mapper.readValue[DtabStoreConfig](yaml).asInstanceOf[ConsulConfig]
+    assert(consul.host.value == "consul.local")
+    assert(consul.port.value == Port(80))
+    assert(consul.pathPrefix == Some(Path.read("/foo/bar")))
+    assert(consul.token == Some("some-token"))
   }
 
 }